requirements.txt
sounddevice
numpy
scipy
mlx-whisper
anthropic
pynput
python-dotenv
